ODS-111 : Feat : ajout composant pour bouton lancement analyse Intégration bouton dans page d'accueil et page de résultat Modification architecture des fichiers / dossiers
