+++ src/components/accueil/BlocRechercheParPpn.vue
