+++ src/components/accueil/BlocTypeAnalyse.vue
