+++ src/components/resultats/BlocRappelTypeAnalyse.vue
